Quell warnings.
Address warnings about unsigned vs signed comparisons.

+++ tests/auto/imports/qcapsulemesh/tst_qcapsulemesh.cpp
@@ -157,7 +157,7 @@ void tst_QCapsuleMesh::testGeometry()
         QGL::IndexArray ib = geom.indices();
         for (int i=pSceneNodeLevel2->start(); i<pSceneNodeLevel2->start()+pSceneNodeLevel2->count(); i+=3) {
             unsigned int ii0 = ib.at(i);
-            QVERIFY(ii0<verts.size());
+            QVERIFY(ii0<(unsigned int)verts.size());
             QVector3D vv0 = verts.at(ii0);
             qreal vv0sq = (vv0-QVector3D(0,0,vv0.z())).lengthSquared();
             QVERIFY(fabs(vv0sq-0.25)<0.0001); // vertex is on cylinder
@@ -165,7 +165,7 @@ void tst_QCapsuleMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn0,(vv0-QVector3D(0,0,vv0.z()))))-1.0) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii1 = ib.at(i+1);
-            QVERIFY(ii1<verts.size());
+            QVERIFY(ii1<(unsigned int)verts.size());
             QVector3D vv1 = verts.at(ii1);
             qreal vv1sq = (vv1-QVector3D(0,0,vv1.z())).lengthSquared();
             QVERIFY(fabs(vv1sq-0.25)<0.0001); // vertex is on cylinder
@@ -173,7 +173,7 @@ void tst_QCapsuleMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn1,(vv1-QVector3D(0,0,vv1.z()))))-1.0) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii2 = ib.at(i+2);
-            QVERIFY(ii2<verts.size());
+            QVERIFY(ii2<(unsigned int)verts.size());
             QVector3D vv2 = verts.at(ii2);
             qreal vv2sq = (vv2-QVector3D(0,0,vv2.z())).lengthSquared();
             QVERIFY(fabs(vv2sq-0.25)<0.0001); // vertex is on cylinder
@@ -209,7 +209,7 @@ void tst_QCapsuleMesh::testGeometry()
         QGL::IndexArray ib = geom.indices();
         for (int i=pSceneNodeLevel2->start(); i<pSceneNodeLevel2->start()+pSceneNodeLevel2->count(); i+=3) {
             unsigned int ii0 = ib.at(i);
-            QVERIFY(ii0<verts.size());
+            QVERIFY(ii0<(unsigned int)verts.size());
             QVector3D vv0 = verts.at(ii0);
             qreal vv0sq = vv0.lengthSquared();
             QVERIFY(fabs(vv0sq-0.25)<0.0001); // vertex is on dome
@@ -219,7 +219,7 @@ void tst_QCapsuleMesh::testGeometry()
             QVERIFY(fabs(QVector3D::dotProduct(nn0,vv0)-0.5) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii1 = ib.at(i+1);
-            QVERIFY(ii1<verts.size());
+            QVERIFY(ii1<(unsigned int)verts.size());
             QVector3D vv1 = verts.at(ii1);
             qreal vv1sq = vv1.lengthSquared();
             QVERIFY(fabs(vv1sq-0.25)<0.0001); // vertex is on dome
@@ -229,7 +229,7 @@ void tst_QCapsuleMesh::testGeometry()
             QVERIFY(fabs(QVector3D::dotProduct(nn1,vv1)-0.5) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii2 = ib.at(i+2);
-            QVERIFY(ii2<verts.size());
+            QVERIFY(ii2<(unsigned int)verts.size());
             QVector3D vv2 = verts.at(ii2);
             qreal vv2sq = vv2.lengthSquared();
             QVERIFY(fabs(vv2sq-0.25)<0.0001); // vertex is on cylinder
@@ -277,7 +277,7 @@ void tst_QCapsuleMesh::testGeometry()
         QGL::IndexArray ib = geom.indices();
         for (int i=pSceneNodeLevel2->start(); i<pSceneNodeLevel2->start()+pSceneNodeLevel2->count(); i+=3) {
             unsigned int ii0 = ib.at(i);
-            QVERIFY(ii0<verts.size());
+            QVERIFY(ii0<(unsigned int)verts.size());
             QVector3D vv0 = verts.at(ii0);
             qreal vv0sq = vv0.lengthSquared();
             QVERIFY(fabs(vv0sq-0.25)<0.0001); // vertex is on dome
@@ -287,7 +287,7 @@ void tst_QCapsuleMesh::testGeometry()
             QVERIFY(fabs(QVector3D::dotProduct(nn0,vv0)-0.5) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii1 = ib.at(i+1);
-            QVERIFY(ii1<verts.size());
+            QVERIFY(ii1<(unsigned int)verts.size());
             QVector3D vv1 = verts.at(ii1);
             qreal vv1sq = vv1.lengthSquared();
             QVERIFY(fabs(vv1sq-0.25)<0.0001); // vertex is on dome
@@ -297,7 +297,7 @@ void tst_QCapsuleMesh::testGeometry()
             QVERIFY(fabs(QVector3D::dotProduct(nn1,vv1)-0.5) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii2 = ib.at(i+2);
-            QVERIFY(ii2<verts.size());
+            QVERIFY(ii2<(unsigned int)verts.size());
             QVector3D vv2 = verts.at(ii2);
             qreal vv2sq = vv2.lengthSquared();
             QVERIFY(fabs(vv2sq-0.25)<0.0001); // vertex is on cylinder
