+++ tests/auto/imports/qcylindermesh/tst_qcylindermesh.cpp
@@ -133,7 +133,7 @@ void tst_QCylinderMesh::testGeometry()
         QGL::IndexArray ib = geom.indices();
         for (int i=pSceneNodeLevel2->start(); i<pSceneNodeLevel2->start()+pSceneNodeLevel2->count(); i+=3) {
             unsigned int ii0 = ib.at(i);
-            QVERIFY(ii0<verts.size());
+            QVERIFY(ii0<(unsigned int)verts.size());
             QVector3D vv0 = verts.at(ii0);
             qreal vv0sq = (vv0-QVector3D(0,0,0.5)).lengthSquared();
             QVERIFY(fabs(vv0sq)<0.0001 || fabs(vv0sq-1.0)<0.0001); // vertex is either in center or on circle
@@ -141,7 +141,7 @@ void tst_QCylinderMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn0,QVector3D(0,0,1)))-1.0) < 0.0001); // vertex normal == {0,0,1}
 
             unsigned int ii1 = ib.at(i+1);
-            QVERIFY(ii1<verts.size());
+            QVERIFY(ii1<(unsigned int)verts.size());
             QVector3D vv1 = verts.at(ii1);
             qreal vv1sq = (vv1-QVector3D(0,0,0.5)).lengthSquared();
             QVERIFY(fabs(vv1sq)<0.0001 || fabs(vv1sq-1.0)<0.0001); // vertex is either in center or on circle
@@ -149,7 +149,7 @@ void tst_QCylinderMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn1,QVector3D(0,0,1)))-1.0) < 0.0001); // vertex normal == {0,0,1}
 
             unsigned int ii2 = ib.at(i+2);
-            QVERIFY(ii2<verts.size());
+            QVERIFY(ii2<(unsigned int)verts.size());
             QVector3D vv2 = verts.at(ii2);
             qreal vv2sq = (vv2-QVector3D(0,0,0.5)).lengthSquared();
             QVERIFY(fabs(vv2sq)<0.0001 || fabs(vv2sq-1.0)<0.0001); // vertex is either in center or on circle
@@ -173,7 +173,7 @@ void tst_QCylinderMesh::testGeometry()
         QGL::IndexArray ib = geom.indices();
         for (int i=pSceneNodeLevel2->start(); i<pSceneNodeLevel2->start()+pSceneNodeLevel2->count(); i+=3) {
             unsigned int ii0 = ib.at(i);
-            QVERIFY(ii0<verts.size());
+            QVERIFY(ii0<(unsigned int)verts.size());
             QVector3D vv0 = verts.at(ii0);
             qreal vv0sq = (vv0-QVector3D(0,0,vv0.z())).lengthSquared();
             QVERIFY(fabs(vv0sq-1.0)<0.0001); // vertex is on cylinder
@@ -181,7 +181,7 @@ void tst_QCylinderMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn0,(vv0-QVector3D(0,0,vv0.z()))))-1.0) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii1 = ib.at(i+1);
-            QVERIFY(ii1<verts.size());
+            QVERIFY(ii1<(unsigned int)verts.size());
             QVector3D vv1 = verts.at(ii1);
             qreal vv1sq = (vv1-QVector3D(0,0,vv1.z())).lengthSquared();
             QVERIFY(fabs(vv1sq-1.0)<0.0001); // vertex is on cylinder
@@ -189,7 +189,7 @@ void tst_QCylinderMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn1,(vv1-QVector3D(0,0,vv1.z()))))-1.0) < 0.0001); // vertex normal points to the same direction as vertex
 
             unsigned int ii2 = ib.at(i+2);
-            QVERIFY(ii2<verts.size());
+            QVERIFY(ii2<(unsigned int)verts.size());
             QVector3D vv2 = verts.at(ii2);
             qreal vv2sq = (vv2-QVector3D(0,0,vv2.z())).lengthSquared();
             QVERIFY(fabs(vv2sq-1.0)<0.0001); // vertex is on cylinder
@@ -214,7 +214,7 @@ void tst_QCylinderMesh::testGeometry()
         QGL::IndexArray ib = geom.indices();
         for (int i=pSceneNodeLevel2->start(); i<pSceneNodeLevel2->start()+pSceneNodeLevel2->count(); i+=3) {
             unsigned int ii0 = ib.at(i);
-            QVERIFY(ii0<verts.size());
+            QVERIFY(ii0<(unsigned int)verts.size());
             QVector3D vv0 = verts.at(ii0);
             qreal vv0sq = (vv0-QVector3D(0,0,-0.5)).lengthSquared();
             QVERIFY(fabs(vv0sq)<0.0001 || fabs(vv0sq-1.0)<0.0001); // vertex is either in center or on circle
@@ -222,7 +222,7 @@ void tst_QCylinderMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn0,QVector3D(0,0,-1)))-1.0) < 0.0001); // vertex normal == {0,0,-1}
 
             unsigned int ii1 = ib.at(i+1);
-            QVERIFY(ii1<verts.size());
+            QVERIFY(ii1<(unsigned int)verts.size());
             QVector3D vv1 = verts.at(ii1);
             qreal vv1sq = (vv1-QVector3D(0,0,-0.5)).lengthSquared();
             QVERIFY(fabs(vv1sq)<0.0001 || fabs(vv1sq-1.0)<0.0001); // vertex is either in center or on circle
@@ -230,7 +230,7 @@ void tst_QCylinderMesh::testGeometry()
             QVERIFY((fabs(QVector3D::dotProduct(nn1,QVector3D(0,0,-1)))-1.0) < 0.0001); // vertex normal == {0,0,-1}
 
             unsigned int ii2 = ib.at(i+2);
-            QVERIFY(ii2<verts.size());
+            QVERIFY(ii2<(unsigned int)verts.size());
             QVector3D vv2 = verts.at(ii2);
             qreal vv2sq = (vv2-QVector3D(0,0,-0.5)).lengthSquared();
             QVERIFY(fabs(vv2sq)<0.0001 || fabs(vv2sq-1.0)<0.0001); // vertex is either in center or on circle
